+++ src/django/api/reports/facility_lists_uploaded.sql
@@ -1,0 +1,10 @@
+SELECT
+  CAST (date_part('month', l.created_at) AS int) AS month,
+  CASE WHEN u.email LIKE '%openapparel.org%' THEN 'y' ELSE 'n' END AS is_public_list,
+  COUNT(*) AS list_count
+FROM api_facilitylist l
+JOIN api_contributor c ON l.contributor_id = c.id
+JOIN api_user u ON u.id = c.admin_id
+WHERE date_part('month', l.created_at) != date_part('month', now())
+GROUP BY date_part('month', l.created_at), is_public_list
+ORDER BY date_part('month', l.created_at), is_public_list;